+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Adds new JSON field `path` to args, parameters and aliases (#244)
 * Fixes for changed JSON format of function and method returns (#245)
 * Fixes for changed JSON format of parent traits (#246)
+* Adds new JSON field `traits` to parameters (#247)
 
 ## [[v0.11.9]](https://github.com/mlange-42/modo/compare/v0.11.8...v0.11.9)